finnie-ai_finance_assistant/requirements.txt
langchain
langgraph
chromadb
openai
streamlit
pandas
plotly
yfinance
pytest
python-dotenv
pyyaml
tiktoken
numpy
langchain-openai
langchain-chroma